Chore MindMap 컴포넌트 MindMapWrapper, MindMap으로 분리 mindmap만 나오는 부분 로직이 길어질 수 있을 것 같아 태근님 말대로 분리
