+++ .eslintrc.json
@@ -8,7 +8,7 @@
   "rules": {
     "react/jsx-props-no-spreading": "off",
     "react/forbid-prop-types": [
-      true,
+      1,
       {
         "forbid": ["any"],
         "checkContextTypes": true,
